pyconsumerrtsp/requirements.txt
opencv_python==3.4.3.18
kafka_python==1.4.3
python-dotenv==0.9.1
